+++ order/src/test/java/com/food/order/PaymentDoTest.java
@@ -15,7 +15,7 @@ import com.food.order.stubrepository.StubOrderService;
 import com.food.order.stubrepository.StubPaymentLogService;
 import com.food.order.stubrepository.StubPaymentService;
 import com.food.order.stubrepository.StubPointService;
-import com.food.order.temp.DefaultPayService;
+import com.food.order.business.DefaultPayService;
 import com.food.common.payment.business.external.PayService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
